backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
bcrypt==4.1.2
python-socketio==5.10.0
eventlet==0.33.3
python-engineio==4.8.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
